+++ os/various/MAG01A.h
@@ -8,4 +8,4 @@ Created by Ondra Sestak 2013
 #define MAG01A_AXIS 0x03
 
 void magInit(I2CDriver *i2c_drv, uint8_t addr);
-void magRead (I2CDriver *i2c_drv, uint8_t addr, uint16_t *mag_x, uint16_t *mag_y, uint16_t *mag_z);
+void magRead (I2CDriver *i2c_drv, uint8_t addr, float *mag_x, float *mag_y, float *mag_z);
